# api/openapi.yaml
openapi: 3.0.0
...
servers:
  - description: Lambda prod server
    url: https://peqalf7v96.execute-api.us-east-1.amazonaws.com/Prod
info:
  description: Vote for your lunch.
  version: "1.0.0"
  title: Simple Lunchtime Voter API
  contact:
    email: tim@lunchtimevoter.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
  /createLunchVote:
    post:
      summary: Creates a new lunch vote
      operationId: createLunchVote
      description: Creates a new lunch vote, returns the GUID for that lunch vote.
      requestBody:
        description: Create a new vote with a particular name at a particular time.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoteCreationRequestObject'
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
...
components:
  schemas:
...
    VoteCreationRequestObject:
      type: object
      required:
        - voteTime
      properties:
        voteTime:
          type: string
          description: The date/time (Pacific Time) to count the votes, eg "12:00" or "1/1/2020 11:30"
        voteName:
          type: string
          description: The name of your lunchtime vote, eg "Tuesday" or "Bob's Going Away Party"
...