# server/src/gen/java/io/swagger/model/VoteCreationRequestObject.java
/*
 * Simple Lunchtime Voter API
 * Vote for your lunch.
 *
 * OpenAPI spec version: 1.0.0
 * Contact: tim@lunchtimevoter.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

package io.swagger.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.v3.oas.annotations.media.Schema;
import javax.validation.constraints.*;

/**
 * VoteCreationRequestObject
 */
@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.JavaJerseyServerCodegen", date = "2019-11-30T17:46:47.852Z[GMT]")public class VoteCreationRequestObject   {
  @JsonProperty("voteName")
  private String voteName = null;

  @JsonProperty("name")
  private String name = null;

  public VoteCreationRequestObject voteName(String voteName) {
    this.voteName = voteName;
    return this;
  }

  /**
   * The date/time (Pacific Time) to count the votes, eg \&quot;12:00\&quot; or \&quot;1/1/2020 11:30\&quot;
   * @return voteName
   **/
  @JsonProperty("voteName")
  @Schema(description = "The date/time (Pacific Time) to count the votes, eg \"12:00\" or \"1/1/2020 11:30\"")
  public String getVoteName() {
    return voteName;
  }

  public void setVoteName(String voteName) {
    this.voteName = voteName;
  }

  public VoteCreationRequestObject name(String name) {
    this.name = name;
    return this;
  }

  /**
   * The name of your lunchtime vote, eg \&quot;Tuesday\&quot; or \&quot;Bob&#x27;s Going Away Party\&quot;
   * @return name
   **/
  @JsonProperty("name")
  @Schema(description = "The name of your lunchtime vote, eg \"Tuesday\" or \"Bob's Going Away Party\"")
  public String getName() {
    return name;
  }

  public void setName(String name) {
    this.name = name;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    VoteCreationRequestObject voteCreationRequestObject = (VoteCreationRequestObject) o;
    return Objects.equals(this.voteName, voteCreationRequestObject.voteName) &&
        Objects.equals(this.name, voteCreationRequestObject.name);
  }

  @Override
  public int hashCode() {
    return Objects.hash(voteName, name);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class VoteCreationRequestObject {\n");
    
    sb.append("    voteName: ").append(toIndentedString(voteName)).append("\n");
    sb.append("    name: ").append(toIndentedString(name)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}